+++ apps/bublys-os/app/world-line/WorldLine/domain/World.ts
@@ -2,22 +2,48 @@ import { Counter } from './Counter';
 
 /**
  * World クラス
- * 各世界の状態を管理する
+ * 世界の状態を表現し、世界線の概念を含む
  */
 export class World {
   public readonly worldId: string;
+  public readonly parentWorldId: string | null;
   public readonly counter: Counter;
+  public readonly apexWorldLineId: string;
 
-  constructor(worldId: string, counter: Counter = new Counter()) {
+  constructor(
+    worldId: string,
+    parentWorldId: string | null = null,
+    counter: Counter = new Counter(),
+    apexWorldLineId: string = ''
+  ) {
     this.worldId = worldId;
+    this.parentWorldId = parentWorldId;
     this.counter = counter;
+    this.apexWorldLineId = apexWorldLineId || worldId; // デフォルトは自身のID
   }
 
   /**
-   * カウンターを更新した新しいWorldを作成
+   * カウンターを更新した新しい世界を作成
    */
   public updateCounter(newCounter: Counter): World {
-    return new World(this.worldId, newCounter);
+    return new World(
+      crypto.randomUUID(),
+      this.worldId,
+      newCounter,
+      this.apexWorldLineId
+    );
+  }
+
+  /**
+   * 現在の世界線IDを更新した新しい世界を作成
+   */
+  public updateCurrentWorldLineId(newWorldLineId: string): World {
+    return new World(
+      this.worldId,
+      this.parentWorldId,
+      this.counter,
+      newWorldLineId
+    );
   }
 
   /**
@@ -26,7 +52,9 @@ export class World {
   public toJson(): object {
     return {
       worldId: this.worldId,
+      parentWorldId: this.parentWorldId,
       counter: this.counter.toJson(),
+      apexWorldLineId: this.apexWorldLineId,
     };
   }
 
@@ -35,8 +63,10 @@ export class World {
    */
   public static fromJson(json: any): World {
     return new World(
-      json.worldId || '',
-      Counter.fromJson(json.counter || {})
+      json.worldId || json.commitId || '', // 後方互換性のためcommitIdもサポート
+      json.parentWorldId || json.parentCommitId || null,
+      Counter.fromJson(json.counter || {}),
+      json.apexWorldLineId || json.worldId || json.commitId || ''
     );
   }
 }
