--- apps/bublys-os/app/world-line/WorldLineGit/ui/CounterView.tsx
@@ -1,70 +1,0 @@
-'use client';
-import React from 'react';
-import { Counter } from '../domain/Counter';
-
-interface CounterViewProps {
-  counter: Counter;
-  onCounterChange: (newCounter: Counter) => void;
-}
-
-export function CounterView({ counter, onCounterChange }: CounterViewProps) {
-  const handleCountUp = () => {
-    onCounterChange(counter.countUp());
-  };
-
-  const handleCountDown = () => {
-    onCounterChange(counter.countDown());
-  };
-
-  return (
-    <div style={{ 
-      display: 'flex', 
-      alignItems: 'center', 
-      justifyContent: 'center', 
-      gap: '1rem',
-      padding: '1rem',
-      backgroundColor: '#f5f5f5',
-      borderRadius: '8px',
-      margin: '1rem 0'
-    }}>
-      <button 
-        onClick={handleCountDown}
-        style={{
-          padding: '0.5rem 1rem',
-          fontSize: '1.2rem',
-          backgroundColor: '#ff5722',
-          color: 'white',
-          border: 'none',
-          borderRadius: '4px',
-          cursor: 'pointer'
-        }}
-      >
-        -
-      </button>
-      
-      <span style={{ 
-        fontSize: '2rem', 
-        fontWeight: 'bold',
-        minWidth: '3rem',
-        textAlign: 'center'
-      }}>
-        {counter.value}
-      </span>
-      
-      <button 
-        onClick={handleCountUp}
-        style={{
-          padding: '0.5rem 1rem',
-          fontSize: '1.2rem',
-          backgroundColor: '#4caf50',
-          color: 'white',
-          border: 'none',
-          borderRadius: '4px',
-          cursor: 'pointer'
-        }}
-      >
-        +
-      </button>
-    </div>
-  );
-}
